src/mlproject/utilis.py
import os 
import sys
from src.mlproject.exception import CustomException
from src.mlproject.logger import logging
import pandas as pd
from dotenv import load_dotenv
import pymysql
import pickle
import numpy as np
# from src.mlproject.utilis import read_sql_data, save_object


load_dotenv()

host=os.getenv("host")
user=os.getenv("user")
passw=os.getenv("password")
db=os.getenv('db')

def read_sql_data():
    logging.info("Reading SQL database started")
    try:
        mydb=pymysql.connect(
            host=host,
            user=user,
            password=passw,
            db=db
            )
        logging.info("conncection Established",mydb)
        df=pd.read_sql_query('Select * from students',mydb)
        print(df.head())

        return df 


    except Exception as ex:
        raise CustomException(ex)
    
def save_object(file_path, obj):
    try:
        dir_path = os.path.dirname(file_path)

        os.makedirs(dir_path, exist_ok=True)

        with open(file_path, "wb") as file_obj:
            pickle.dump(obj, file_obj)
        
    except Exception as e:
        raise CustomException(e, sys)

    